Some small refactoring

+++ Travian Manager/TMVillage.h
@@ -23,16 +23,16 @@
 @property(nonatomic, strong) NSArray *constructions; // Construction list
 @property(nonatomic, strong) NSString *name; // Village's name
 @property(nonatomic, strong) NSString *urlPart; // Village ID required for switching villages in url
-@property(assign) int population; // Village population
-@property(assign) int loyalty; // Village loyalty
-@property(assign) unsigned int warehouse; // Max storage of resources (other than wheat)
-@property(assign) unsigned int granary; // Max storage wheat
-@property(assign) unsigned int consumption; // Consuming / hour
-@property(assign) int x; // Location X
-@property(assign) int y; // Location Y
+@property(assign) NSUInteger population; // Village population
+@property(assign) NSInteger loyalty; // Village loyalty
+@property(assign) NSUInteger warehouse; // Max storage of resources (other than wheat)
+@property(assign) NSUInteger granary; // Max storage wheat
+@property(assign) NSUInteger consumption; // Consuming / hour
+@property(assign) NSInteger x; // Location X
+@property(assign) NSInteger y; // Location Y
 @property(nonatomic, strong) TMFarmList *farmList; // Farm List object
 
-@property(assign) bool hasDownloaded; // Has downloaded property (used with loadallatonce property of TMSettings)
+@property(assign) bool hasDownloaded; // Has downloaded property (used with load all at once property of TMSettings)
 
 - (void)setAccountParent:(TMAccount *)newParent;
 - (TMAccount *)getParent;
